# src/spec/swagger.json
{"openapi":"3.0.0","paths":{"/health-check":{"get":{"operationId":"AppHealthCheck","parameters":[],"responses":{"200":{"description":""}}}},"/prices/allTokens":{"get":{"operationId":"PricesGetAllTokens","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}},"tags":["prices"]}},"/prices/allPorts":{"get":{"operationId":"PricesGetAllPorts","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"number"}}}}}},"tags":["prices"]}},"/prices/{port}/{token}/{intervals}":{"post":{"operationId":"PricesAdd","parameters":[{"name":"port","required":true,"in":"path","schema":{"type":"number"}},{"name":"token","required":true,"in":"path","schema":{"type":"string"}},{"name":"intervals","required":true,"in":"path","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"201":{"description":""}},"tags":["prices"]}},"/prices/{token}":{"get":{"operationId":"PricesGetPort","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["prices"]},"delete":{"operationId":"PricesRemove","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["prices"]}},"/identity/config":{"get":{"operationId":"IdentityGetConfig","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigModel"}}}}},"tags":["identity"]}},"/signals/allSignals":{"get":{"operationId":"SignalsGetAllSignals","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}},"tags":["signals"]}},"/signals/allPorts":{"get":{"operationId":"SignalsGetAllPorts","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"number"}}}}}},"tags":["signals"]}},"/signals/{id}":{"post":{"operationId":"SignalsAdd","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigSignalInputModel"}}}},"responses":{"201":{"description":""}},"tags":["signals"]},"delete":{"operationId":"SignalsRemove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["signals"]}},"/signals/{id}/port":{"get":{"operationId":"SignalsGetPort","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["signals"]}},"/signals/{id}/tokens":{"get":{"operationId":"SignalsGetTokens","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}},"tags":["signals"]}},"/transactions/wallet/free":{"get":{"operationId":"TransactionGetWalletFree","parameters":[],"responses":{"200":{"description":""}},"tags":["transaction"]}},"/transactions/wallet/locked":{"get":{"operationId":"TransactionGetWalletLocked","parameters":[],"responses":{"200":{"description":""}},"tags":["transaction"]}},"/setups/all":{"get":{"operationId":"TradingSetupsGetAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TradingSetupModel"}}}}}},"tags":["trading"]}},"/setups/{id}":{"get":{"operationId":"TradingSetupsGet","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradingSetupModel"}}}}},"tags":["trading"]},"delete":{"operationId":"TradingSetupsRemove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradingSetupModel"}}}}},"tags":["trading"]}},"/setups/{id}/{startingFirstAmount}/{startingSecondAmount}":{"post":{"operationId":"TradingSetupsAdd","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"startingFirstAmount","required":true,"in":"path","schema":{"type":"string"}},{"name":"startingSecondAmount","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradingSetupConfigModel"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradingSetupModel"}}}}},"tags":["trading"]}}},"info":{"title":"TradingBot","description":"","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"/"}],"components":{"schemas":{"StringObject":{"type":"object","properties":{"value":{"type":"string"}},"required":["value"]},"ConfigModel":{"type":"object","properties":{"intervals":{"default":[],"type":"array","items":{"type":"string"}}},"required":["intervals"]},"ConfigSignalInputModel":{"type":"object","properties":{"port":{"type":"number"},"tokens":{"type":"array","items":{"type":"string"}},"intervals":{"type":"array","items":{"type":"string"}}},"required":["port","tokens","intervals"]},"TradingSetupStatusType":{"type":"string","enum":["INITIAL","RUNNING","TERMINATED"]},"TradingTakeProfitTrailingStopConfigModel":{"type":"object","properties":{"deltaPercentage":{"type":"number","default":0.05},"hardLimitPercentage":{"type":"number"}},"required":["deltaPercentage"]},"TradingTakeProfitConfigModel":{"type":"object","properties":{"percentage":{"type":"number"},"trailingStop":{"$ref":"#/components/schemas/TradingTakeProfitTrailingStopConfigModel"}},"required":["percentage"]},"TradingStopLossConfigModel":{"type":"object","properties":{"percentage":{"type":"number"}},"required":["percentage"]},"TradingSetupConfigModel":{"type":"object","properties":{"firstToken":{"type":"string"},"secondToken":{"type":"string"},"interval":{"type":"string"},"signal":{"type":"string"},"terminationPercentageLoss":{"type":"number"},"takeProfit":{"$ref":"#/components/schemas/TradingTakeProfitConfigModel"},"stopLoss":{"$ref":"#/components/schemas/TradingStopLossConfigModel"},"useLimitOrders":{"type":"boolean","default":false},"limitOrderBuyOffset":{"type":"number","default":0},"limitOrderSellOffset":{"type":"number","default":0},"limitOrderCancelDueToChecksElapsed":{"type":"number","default":10000},"limitOrderCancelDueToTimeElapsed":{"type":"number"},"limitOrderCancelDueToPriceDivergence":{"type":"string"}},"required":["firstToken","secondToken","interval","signal","useLimitOrders","limitOrderBuyOffset","limitOrderSellOffset","limitOrderCancelDueToChecksElapsed"]},"TradingTransactionModel":{"type":"object","properties":{"buy":{"type":"boolean"},"firstToken":{"type":"string"},"secondToken":{"type":"string"},"firstAmount":{"type":"string"},"secondAmount":{"type":"string"},"priceAmount":{"type":"string"},"transactionId":{"type":"string"},"complete":{"type":"boolean","default":false},"checks":{"type":"number","default":0},"firstUpdateTimestamp":{"type":"number","default":0},"lastUpdateTimestamp":{"type":"number","default":0}},"required":["buy","firstToken","secondToken","firstAmount","secondAmount","priceAmount","transactionId","complete","checks","firstUpdateTimestamp","lastUpdateTimestamp"]},"TradingSetupModel":{"type":"object","properties":{"id":{"type":"string"},"status":{"$ref":"#/components/schemas/TradingSetupStatusType"},"config":{"$ref":"#/components/schemas/TradingSetupConfigModel"},"startingFirstAmount":{"type":"string"},"startingSecondAmount":{"type":"string"},"firstAmount":{"type":"string"},"secondAmount":{"type":"string"},"lastUpdate":{"type":"integer","default":1697476142852,"format":"int64"},"currentPriceAmount":{"type":"string","default":"0"},"lowestPriceAmount":{"type":"string","default":"99999999999"},"highestPriceAmount":{"type":"string","default":"0"},"tradeEntryPriceAmount":{"type":"string","default":"0"},"tradeLowestPriceAmount":{"type":"string","default":"99999999999"},"tradeHighestPriceAmount":{"type":"string","default":"0"},"transactions":{"default":[],"type":"array","items":{"$ref":"#/components/schemas/TradingTransactionModel"}},"openTransactions":{"type":"object"}},"required":["id","status","config","startingFirstAmount","startingSecondAmount","firstAmount","secondAmount","lastUpdate","currentPriceAmount","lowestPriceAmount","highestPriceAmount","tradeEntryPriceAmount","tradeLowestPriceAmount","tradeHighestPriceAmount","transactions","openTransactions"]}}}}